Add some package-related TODOs Move I02_1FgsParams
